Add an option to specify the button to continue
jscal asks user to move axis to a position and push any button.
This causes a problem when a trigger of a gamepad is both axis and a
button. In this case when user pulls the trigger, a button event is sent
and jscal continues before user has had a change to move the trigger to
the requested position. The problem appears with ZEROPLUS P4 Wired Gamepad.

This patch adds a -b option to jscal to choose one button to use as
the continue button.

Signed-off-by: Stephen Kitt <steve@sk2.org>