Support for ELO touchscreen using General Touch ST6001S controller
This is a patch in order to ask the Linux kernel to support a variant
of the ELO touchscreen that requires a different initialisation string
in order to function.

The patch allows inputattach to select a different ELO id
parameter (4).

The kernel patch for the Linux change is not yet merged but the code
for it can be seen in this post to the relevant mailing list:
https://lore.kernel.org/linux-input/83a934eb-e1ee-fd79-33ee-76413be2e6ea@virgin.net/T/#m6caed1d2c1bb262440cd7950207c7fa67b17edb1